Adding the capability to skip certain disks by container-type
This adds the possibility to skip disks that are in a certain
container type. The types can be specified by the switch
--skip-container-type (can be specified multiple times) and
are ignored when the check for multipathing runs.

The background for this change is that the NetAPP mediator
adds virtual disks without multipathing to the NetAPP head. Those
disks are part of the container-type mediator and now can be
ignored by adding the parameter on script execution.

However, there may be other types of disks that do not support
multipathing, so a parameter makes the solution more versatile.